requirements.txt
# requirements.txt
# Dependencies required for the Family Task Bot with Python 3.13 compatibility
# Last Modified: 2024-11-01
# Using Python 3.12

# Core dependencies
python-telegram-bot>=21.0    # Updated version for Python 3.13 compatibility
# schedule==1.2.2              # Scheduling library for daily reminders
python-dotenv==1.0.1         # Load environment variables from .env file

# Python 3.13 compatibility dependencies
# httpx>=0.27.0                # HTTP client (used by python-telegram-bot and ultra_minimal_bot)
pytz>=2023.3                 # Timezone library required by APScheduler
tzlocal==2.1                 # Compatible version for timezone handling

# Optional: For ultra minimal implementation
# Uncomment if using ultra_minimal_bot.py instead of the main implementation
# httpx>=0.27.0              # HTTP client for direct Telegram API calls